+++ modules/genealogy-discovery-livewire/src/DiscoveryLivewireServiceProvider.php
@@ -14,8 +14,10 @@ final class DiscoveryLivewireServiceProvider extends ServiceProvider
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'genealogy-discovery-livewire');
         Livewire::component('genealogy-discovery-list', DiscoveryMatchList::class);
         Livewire::component('genealogy-discovery-search', DiscoverySearch::class);
+        Livewire::component('genealogy-discovery-external-search', ExternalDiscoverySearch::class);
         Livewire::component('module-genealogy-discovery::discovery-list', DiscoveryMatchList::class);
         Livewire::component('module-genealogy-discovery::discovery-search', DiscoverySearch::class);
+        Livewire::component('module-genealogy-discovery::external-discovery-search', ExternalDiscoverySearch::class);
     }
 }